Extract env interpolation into its own module
Move the config env-variable interpolation out of system_config.rs into
a dedicated config_parsing::env_interpolation module, renamed from the
generic `interpolation` for clarity.
